+++ src/fn/initialise-project-layout.js
@@ -20,7 +20,7 @@ const LAYOUT = {
   '.eslintrc': `{
   "env": {
     "node": true,
-    "es2022": true,
+    "es2022": true
   },
   "parserOptions": {
     "ecmaVersion": 2022